[P216T10A1]

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -12,7 +12,7 @@ int _atoi(char *s)
 	int found = 0;
 	char *number_head = find_number_head(s, &found);
 	int sign = positive_negative_none(number_head);
-	int number = 0;
+	long long number = 0;
 
 	while (*number_head != '\0' && ((*number_head >= 48 && *number_head <= 57)))
 	{
